Refactor to use config pattern
